+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Make the engine faster.
 
+### Fixed
+
+- Hnefatafl Text Protocol: generate_move
+
 ## [4.1.1] - 2025-09-23
 
 ### Added