+++ CHANGELOG.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- hnefatafl-client: use postcard instead of ron for the user config file.
+- hnefatafl-client: use postcard instead of ron for the user config file
+  (breaking, but it is not important).
 - Make the engine faster.
 
 ### Fixed