+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.0] - 2025-09-23
+
 ### Added
 
 - hnefatafl-client: the option to save your password.
@@ -470,7 +472,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hnefatafl-client: Make users and games scrollable.
 - Make all the features dependencies.
 
-[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v4.1.1...main
+[unreleased]: https://github.com/dcampbell24/hnefatafl/compare/v4.2.0...main
+[4.2.0]: https://github.com/dcampbell24/hnefatafl/compare/v4.1.1...v4.2.0
 [4.1.1]: https://github.com/dcampbell24/hnefatafl/compare/v4.1.0...v4.1.1
 [4.1.0]: https://github.com/dcampbell24/hnefatafl/compare/v4.0.0...v4.1.0
 [4.0.0]: https://github.com/dcampbell24/hnefatafl/compare/v3.0.1...v4.0.0